+++ libraries/SX126x-Arduino/examples/LORAWAN-OTAA-eora-s3-900tb/src/utilities.h
@@ -1,0 +1,54 @@
+#pragma once
+
+#define UNUSE_PIN                   (-1)
+
+#define GPIO_PIN_16                 16
+#define GPIO_PIN_15                 15
+#define GPIO_PIN_RST                UNUSE_PIN
+#define GPIO_PIN_12                 12
+#define GPIO_PIN_48                 48
+#define GPIO_PIN_47                 47
+#define GPIO_PIN_35                 35
+#define GPIO_PIN_00                 BUTTON_PIN
+#define GPIO_PIN_36                 36
+#define GPIO_PIN_37                 BOARD_LED
+
+#define GPIO_PIN_38                 38
+#define GPIO_PIN_39                 39
+#define GPIO_PIN_40                 40
+#define GPIO_PIN_41                 41
+#define GPIO_PIN_45                 45
+#define GPIO_PIN_46                 46
+#define GPIO_PIN_42                 42
+
+#define I2C_SDA                     18
+#define I2C_SCL                     17
+#define OLED_RST                    UNUSE_PIN
+
+#define RADIO_SCLK_PIN              5
+#define RADIO_MISO_PIN              3
+#define RADIO_MOSI_PIN              6
+#define RADIO_CS_PIN                7
+#define RADIO_DIO1_PIN              33
+#define RADIO_BUSY_PIN              34
+#define RADIO_RST_PIN               8
+
+#define SDCARD_MOSI                 11
+#define SDCARD_MISO                 2
+#define SDCARD_SCLK                 14
+#define SDCARD_CS                   13
+
+#define BOARD_LED                   37
+#define LED_BUILTIN                 BOARD_LED
+#define LED_ON                      HIGH
+#define LED_OFF                     LOW
+
+#define BAT_ADC_PIN                 1
+#define BUTTON_PIN                  0
+
+#define HAS_SDCARD
+#define HAS_DISPLAY
+
+
+
+