+++ libraries/SX126x-Arduino/src/boards/mcu/board.h
@@ -77,7 +77,7 @@ struct hw_config
 	RadioTcxoCtrlVoltage_t TCXO_CTRL_VOLTAGE = TCXO_CTRL_3_3V;
 };
 
-#ifdef ARDUINO_ARCH_RP2040
+#if defined ARDUINO_ARCH_RP2040 && not defined ARDUINO_RAKWIRELESS_RAK11300
 #include <mbed.h>
 #include <rtos.h>
 // Wake up LoRa event handler on RP2040
@@ -108,9 +108,13 @@ uint32_t lora_isp4520_init(int chipType);
  */
 uint32_t lora_rak4630_init(void);
 
-/**@brief Initializes the RAK11300 board peripherals.
+/**@brief Initializes the ARDUINO_RAKWIRELESS_RAK11300 board peripherals.
  */
 uint32_t lora_rak11300_init(void);
+
+/**@brief Initializes the RAK3112 board peripherals.
+ */
+uint32_t lora_rak3112_init(void);
 
 /**@brief Initializes the RAK13300 board peripherals.
  */