+++ libraries/SX126x-Arduino/src/mac/LoRaMacHelper.cpp
@@ -24,6 +24,8 @@
 #include "system/utilities.h"
 
 #include "mac/LoRaMacTest.h"
+#include "mac/LoRaMac.h"
+extern LoRaMacParams_t LoRaMacParams;
 
 uint16_t ChannelsMask[6];
 uint16_t ChannelsDefaultMask[6];
@@ -36,6 +38,7 @@ LoRaMacRegion_t region;
 bool _otaa = false;
 
 bool _dutyCycleEnabled = false;
+extern bool PublicNetwork;
 
 bool lmh_mac_is_busy = false;
 
@@ -380,7 +383,8 @@ static void McpsConfirm(McpsConfirm_t *mcpsConfirm)
 		// Check Datarate
 		// Check TxPower
 		// Report unconfirmed TX finished
-		if (!statusOk) LOG_LIB("LMH", "Timeout TX + RX finished");
+		if (!statusOk)
+			LOG_LIB("LMH", "Timeout TX + RX finished");
 		if (m_callbacks->lmh_unconf_finished != 0)
 		{
 			m_callbacks->lmh_unconf_finished();
@@ -395,10 +399,16 @@ static void McpsConfirm(McpsConfirm_t *mcpsConfirm)
 		// Check NbTrials
 
 		// Report confirmed TX finished with result
-		if (!statusOk) LOG_LIB("LMH", "Timeout Conf TX finished %s", mcpsConfirm->AckReceived ? "SUCC" : "FAIL");
+		if (!statusOk)
+			LOG_LIB("LMH", "Timeout Conf TX finished %s", mcpsConfirm->AckReceived ? "SUCC" : "FAIL");
 		if (m_callbacks->lmh_conf_result != 0)
 		{
 			m_callbacks->lmh_conf_result(mcpsConfirm->AckReceived);
+			// Workaround, reset MAC state
+			// Workaround for DR reset when ADR is active
+			int8_t preserve_dr = LoRaMacParams.ChannelsDatarate;
+			lmh_reset_mac();
+			LoRaMacParams.ChannelsDatarate = preserve_dr;
 		}
 		break;
 	}
@@ -661,6 +671,8 @@ static void MlmeConfirm(MlmeConfirm_t *mlmeConfirm)
 			if (m_callbacks->lmh_has_joined_failed != NULL)
 			{
 				m_callbacks->lmh_has_joined_failed();
+				// Workaround, reset MAC state
+				lmh_reset_mac();
 			}
 
 			// Join was not successful. Try to join again
@@ -691,7 +703,7 @@ static void MlmeConfirm(MlmeConfirm_t *mlmeConfirm)
 }
 
 lmh_error_status lmh_init(lmh_callback_t *callbacks, lmh_param_t lora_param, bool otaa,
-						  eDeviceClass nodeClass, LoRaMacRegion_t user_region)
+						  eDeviceClass nodeClass, LoRaMacRegion_t user_region, bool region_change)
 {
 	region = (LoRaMacRegion_t)user_region;
 	char strlog1[64];
@@ -705,6 +717,8 @@ lmh_error_status lmh_init(lmh_callback_t *callbacks, lmh_param_t lora_param, boo
 	_otaa = otaa;
 
 	_dutyCycleEnabled = m_param.duty_cycle;
+
+	PublicNetwork = m_param.enable_public_network;
 
 #if (STATIC_DEVICE_EUI != 1)
 	m_callbacks->BoardGetUniqueId(DevEui);
@@ -742,7 +756,7 @@ lmh_error_status lmh_init(lmh_callback_t *callbacks, lmh_param_t lora_param, boo
 	LoRaMacPrimitives.MacMlmeConfirm = MlmeConfirm;
 	LoRaMacCallbacks.GetBatteryLevel = m_callbacks->BoardGetBatteryLevel;
 
-	error_status = LoRaMacInitialization(&LoRaMacPrimitives, &LoRaMacCallbacks, region, nodeClass);
+	error_status = LoRaMacInitialization(&LoRaMacPrimitives, &LoRaMacCallbacks, region, nodeClass, region_change);
 	if (error_status != LORAMAC_STATUS_OK)
 	{
 		return LMH_ERROR;
@@ -994,7 +1008,7 @@ lmh_error_status lmh_send(lmh_app_data_t *app_data, lmh_confirm is_tx_confirmed)
 			// 	(region == LORAMAC_REGION_AS923_4))
 			if (region == LORAMAC_REGION_AS923)
 			{
-				mcpsReq.Req.Confirmed.NbTrials = 1; //8;
+				mcpsReq.Req.Confirmed.NbTrials = 1; // 8;
 			}
 			else
 			{
@@ -1088,7 +1102,7 @@ lmh_error_status lmh_class_request(DeviceClass_t newClass)
 
 		case CLASS_C:
 		{
-			if (currentClass != CLASS_A)
+			if (currentClass != CLASS_C)
 			{
 				Errorstatus = LMH_ERROR;
 			}
@@ -1114,6 +1128,11 @@ lmh_error_status lmh_class_request(DeviceClass_t newClass)
 	return Errorstatus;
 }
 
+/**
+ * @brief Get the device class
+ *
+ * @param currentClass 0 or 2 for Class A or C (Class B is not supported)
+ */
 void lmh_class_get(DeviceClass_t *currentClass)
 {
 	MibRequestConfirm_t mibReq;
@@ -1124,16 +1143,51 @@ void lmh_class_get(DeviceClass_t *currentClass)
 	*currentClass = mibReq.Param.Class;
 }
 
+/**
+ * @brief Get device LoRaWAN address
+ *
+ * @return uint32_t device address
+ */
 uint32_t lmh_getDevAddr(void)
 {
 	return LoRaMacGetOTAADevId();
 }
 
+void lmh_getNwSkey(uint8_t *key)
+{
+	for (int idx = 0; idx < 16; idx++)
+	{
+		key[idx] = LoRaMacNwkSKey[idx];
+	}
+}
+
+void lmh_getAppSkey(uint8_t *key)
+{
+	for (int idx = 0; idx < 16; idx++)
+	{
+		key[idx] = LoRaMacAppSKey[idx];
+	}
+}
+
+/**
+ * @brief Set the AS923 frequency variant
+ *
+ * @param version 1, 2, 3 or 4 for AS923-1 (default), AS923-2, AS923-3 or AS923-4
+ * @return true
+ * @return false
+ */
 bool lmh_setAS923Version(uint8_t version)
 {
 	return RegionAS923SetVersion(version);
 }
 
+/**
+ * @brief Set number of retries for confirmed packages
+ *
+ * @param retries number of retries
+ * @return true if success
+ * @return false if failed (number of retries to small or large)
+ */
 bool lmh_setConfRetries(uint8_t retries)
 {
 	if ((retries > 0) && (retries < 8))
@@ -1144,7 +1198,21 @@ bool lmh_setConfRetries(uint8_t retries)
 	return false;
 }
 
+/**
+ * @brief Get number of retries
+ *
+ * @return uint8_t number of retries
+ */
 uint8_t lmh_getConfRetries(void)
 {
 	return max_ack_retries;
+}
+
+/**
+ * @brief Reset MAC parameters
+ *
+ */
+void lmh_reset_mac(void)
+{
+	ResetMacCounters();
 }