+++ libraries/SX126x-Arduino/CHANGELOG.md
@@ -4,6 +4,45 @@ Arduino library for LoRa communication with Semtech SX126x chips. It is based on
 
 # Release Notes
 
+# V2.0.31 Fix RAK3112
+  - Correct LoRa pin assignment for RAK3112
+  
+# V2.0.30 Fix RP2040 assert issue
+  - Set timer priority correct, thanks to _**@EdisonAgudelo**_
+  - Add RAK3112 support
+
+# V2.0.29 Add read the Syncword function
+  - Add radio function to read the radio syncword uint16_t syncword = Radio.GetSyncWord(void);
+  
+# V2.0.28 Add custom Syncword and Low DataRate optimization
+  - Add radio function to set a custom Syncword Radio.SetCustomSyncWord(uint16_t syncword);
+  - Add radio function to enforce Low DataRate optimization Radio.EnforceLowDRopt(bool enforce);
+  
+# V2.0.27 Access to NWsKey and AppsKey
+  - Add functions to get network session key and app session key after join
+  
+# V2.0.26 Add missing header file
+  - Fix compilation error
+  
+# V2.0.25 Fix DR reset with ADR enabled
+  - With ADR enabled, fix the DR reset to default when confirmed/unconfirmed packets are sent.
+  
+## V2.0.24 Fix P2P RX timeout callback fails
+  - Fix callbacks for P2P RX and TX timeout not being called
+
+## V2.0.23 Add support for the Arduino Pico BSP
+  - Added support for the RAK11300 in [arduino-pico](https://github.com/earlephilhower/arduino-pico) BSP  _**still experimental**_
+
+## V2.0.22 IRQ_HEADER_ERROR should issue RX error, not RX timeout
+  - Change behaviour of IRQ_HEADER_ERROR, thanks to _**@JeromeBriot**_
+  
+## V2.0.21 Some fixes
+  - Fix public/private network always public
+  - Add option to restart MAC stack to change e.g. region without re-init timers
+
+## V2.0.20 Add MAC parameter reset function
+  - Add option to reset the MAC counters
+
 ## V2.0.19 Fix the release mess
   - Sorry released the wrong branch